feat(metrics): refactor VLM metric imports and enhance VieScoreMetric functionality
- Updated import paths for `StatefulVLMMeanScoresMetric` and related functions to reflect the new structure.
- Added `StatefulVLMMeanScoresMetric` to the `__all__` exports in `__init__.py` for better accessibility.
- Removed the deprecated `metric_vlm_base.py` and `viescore_prompts.py` files to streamline the codebase.
- Enhanced the `VieScoreMetric` class with new prompt-building functions for improved clarity and functionality.
- Updated relevant metric files to utilize the new import structure and functions.
