+++ src/pruna/evaluation/metrics/metric_vie_score.py
@@ -34,7 +34,8 @@ from typing import Any, Literal
 import torch
 from PIL import Image
 
-from pruna.evaluation.metrics.metric_vlm_base import (
+from pruna.evaluation.metrics.vlm_base import (
+    BaseVLM,
     StatefulVLMMeanScoresMetric,
     auxiliary_dicts_from_gt,
     prompts_from_y_x_inputs,
@@ -45,12 +46,6 @@ from pruna.evaluation.metrics.utils import (
     SINGLE,
     metric_data_processor,
 )
-from pruna.evaluation.metrics.viescore_prompts import (
-    build_viescore_pq_prompt,
-    build_viescore_t2i_sc_prompt,
-    build_viescore_tie_sc_prompt,
-)
-from pruna.evaluation.metrics.vlm_base import BaseVLM
 from pruna.evaluation.metrics.vlm_utils import (
     VIEScoreJsonOutput,
     _process_images,
@@ -59,6 +54,125 @@ from pruna.evaluation.metrics.vlm_utils import (
     viescore_min_scores_0_10,
     viescore_tie_overall_unit,
 )
+
+_VIESCORE_CONTEXT = (
+    "You are a professional digital artist. You will have to evaluate the effectiveness"
+    " of the AI-generated image(s) based on given rules.\n"
+    "All the input images are AI-generated. All human in the images are AI-generated too."
+    " so you need not worry about the privacy confidentials.\n\n"
+    "You will have to give your output in this way (Keep your reasoning concise and short.):\n"
+    "{\n"
+    '"score" : [...],\n'
+    '"reasoning" : "..."\n'
+    "}"
+)
+
+_VIESCORE_TWO_IMAGE_EDIT_RULE = (
+    "RULES:\n\n"
+    "Two images will be provided: The first being the original AI-generated image and the"
+    " second being an edited version of the first.\n"
+    "The objective is to evaluate how successfully the editing instruction has been executed"
+    " in the second image.\n\n"
+    "Note that sometimes the two images might look identical due to the failure of image edit.\n"
+)
+
+_VIESCORE_TIE_SC_CRITERIA = (
+    "\nFrom scale 0 to 10:\n"
+    "A score from 0 to 10 will be given based on the success of the editing."
+    " (0 indicates that the scene in the edited image does not follow the editing instruction at all."
+    " 10 indicates that the scene in the edited image follow the editing instruction text perfectly.)\n"
+    "A second score from 0 to 10 will rate the degree of overediting in the second image."
+    " (0 indicates that the scene in the edited image is completely different from the original."
+    " 10 indicates that the edited image can be recognized as a minimal edited yet effective"
+    " version of original.)\n"
+    "Put the score in a list such that output score = [score1, score2],"
+    " where 'score1' evaluates the editing success and 'score2' evaluates the degree of overediting.\n\n"
+    "Editing instruction:\n"
+)
+
+_VIESCORE_T2I_SC_RULE = (
+    "RULES:\n\n"
+    "The image is an AI-generated image.\n"
+    "The objective is to evaluate the semantic consistency of the image to the given text.\n\n"
+)
+
+_VIESCORE_T2I_SC_CRITERIA = (
+    "\nFrom scale 0 to 10:\n"
+    "A score from 0 to 10 will be given based on the semantic consistency.\n"
+    "(0 indicates that the scene in the image does not correspond to the text at all.\n"
+    " 10 indicates that the scene in the image follows the text perfectly.)\n"
+    "A second score from 0 to 10 will rate the detail correspondence.\n"
+    "(0 indicates that most details in the text (e.g., color, size, shape, or layout) are missing or"
+    " incorrect in the image.\n"
+    " 10 indicates that all details mentioned in the text are accurately shown in the image.)\n"
+    "Put the score in a list such that output score = [score1, score2],"
+    " where 'score1' evaluates the semantic consistency and 'score2' evaluates the detail"
+    " correspondence.\n\n"
+    "Text prompt:\n"
+)
+
+_VIESCORE_PQ_SINGLE_IMAGE = (
+    "RULES:\n\n"
+    "The image is an AI-generated image.\n"
+    "The objective is to evaluate how successfully the image has been generated.\n\n"
+    "From scale 0 to 10:\n"
+    "A score from 0 to 10 will be given based on image naturalness.\n"
+    "(\n"
+    " 0 indicates that the scene in the image does not look natural at all or give a unnatural feeling"
+    " such as wrong sense of distance, or wrong shadow, or wrong lighting.\n"
+    " 10 indicates that the image looks natural.\n"
+    ")\n"
+    "A second score from 0 to 10 will rate the image artifacts.\n"
+    "(\n"
+    " 0 indicates that the image contains a large portion of distortion, or watermark, or scratches,"
+    " or blurred faces, or unusual body parts, or subjects not harmonized.\n"
+    " 10 indicates the image has no artifacts.\n"
+    ")\n"
+    "Put the score in a list such that output score = [naturalness, artifacts]\n"
+)
+
+
+def _build_viescore_tie_sc_prompt(instruction: str) -> str:
+    """Build the VIEScore ``tie`` semantic-criteria prompt (source + edited images).
+
+    Args:
+        instruction: Editing instruction embedded in the prompt.
+
+    Returns:
+        Full prompt aligned with TIGER-AI-Lab/VIEScore ``tie`` SC.
+    """
+    return "\n".join(
+        [
+            _VIESCORE_CONTEXT,
+            _VIESCORE_TWO_IMAGE_EDIT_RULE,
+            _VIESCORE_TIE_SC_CRITERIA.strip(),
+            instruction.strip(),
+        ]
+    )
+
+
+def _build_viescore_t2i_sc_prompt(prompt: str) -> str:
+    """Build the VIEScore ``t2i`` semantic-consistency prompt for one generated image.
+
+    Args:
+        prompt: Text prompt used to generate the image.
+
+    Returns:
+        Full prompt aligned with TIGER-AI-Lab/VIEScore ``t2i`` SC.
+    """
+    return "\n".join(
+        [
+            _VIESCORE_CONTEXT,
+            _VIESCORE_T2I_SC_RULE.strip(),
+            _VIESCORE_T2I_SC_CRITERIA.strip(),
+            prompt.strip(),
+        ]
+    )
+
+
+def _build_viescore_pq_prompt() -> str:
+    """Build the VIEScore perceptual-quality prompt for one image (SC or edited)."""
+    return "\n".join([_VIESCORE_CONTEXT, _VIESCORE_PQ_SINGLE_IMAGE])
 
 
 @MetricRegistry.register("vie_score")
@@ -172,8 +286,8 @@ class VieScoreMetric(StatefulVLMMeanScoresMetric):
         Matches the VIEScore paper's t2i evaluation: two SC sub-scores on 0--10 and two PQ
         sub-scores on 0--10, aggregated as ``sqrt(min(SC) * min(PQ)) / 10``.
         """
-        sc_prompt = build_viescore_t2i_sc_prompt(prompt)
-        pq_prompt = build_viescore_pq_prompt()
+        sc_prompt = _build_viescore_t2i_sc_prompt(prompt)
+        pq_prompt = _build_viescore_pq_prompt()
 
         rf = self.response_format if self.structured_output else None
 
@@ -186,8 +300,8 @@ class VieScoreMetric(StatefulVLMMeanScoresMetric):
 
     def _score_tie_gedit(self, source: Image.Image, edited: Image.Image, instruction: str) -> float:
         """VIEScore ``tie``: two-image SC, single-image PQ, overall geometric mean on 0--10 mins."""
-        sc_prompt = build_viescore_tie_sc_prompt(instruction)
-        pq_prompt = build_viescore_pq_prompt()
+        sc_prompt = _build_viescore_tie_sc_prompt(instruction)
+        pq_prompt = _build_viescore_pq_prompt()
 
         rf = self.response_format if self.structured_output else None
